refactor: all express middlewares should be defined in the server app
