+++ apps/server/src/main.ts
@@ -14,7 +14,6 @@ import { SyncServer } from "@mp/sync-server";
 import { measureTimeSpan, Ticker, TimeSpan } from "@mp/time";
 import {
   collectDefaultMetrics,
-  createMetricsScrapeMiddleware,
   MetricsGague,
   MetricsHistogram,
   MetricsRegistry,
@@ -32,7 +31,8 @@ import { ClientRegistry } from "./modules/world/ClientRegistry";
 import { createRootRouter } from "./modules/router";
 import { tokenHeaderName } from "./shared";
 import { collectUserMetrics } from "./modules/world/collectUserMetrics";
-import { clientMiddleware } from "./clientMiddleware";
+import { metricsMiddleware } from "./express/metricsMiddleware";
+import { clientMiddleware } from "./express/clientMiddleware";
 
 const logger = new Logger(console);
 
@@ -105,7 +105,7 @@ const expressStaticConfig = {
 const webServer = express()
   .set("trust proxy", opt.trustProxy)
   .use(expressLogger)
-  .use(createMetricsScrapeMiddleware(metrics))
+  .use(metricsMiddleware(metrics))
   .use(createCors({ origin: opt.corsOrigin }))
   .use(opt.publicPath, express.static(opt.publicDir, expressStaticConfig));
